fix: make risc-v builds work
