+++ Cargo.toml
@@ -60,8 +60,11 @@ keywords = ["blockchain", "zksync", "zk", "risc-v"]
 categories = ["cryptography"]
 
 [workspace.dependencies]
-zksync_os_interface = { git = "https://github.com/matter-labs/zksync-os-interface", rev = "8482ed4c8d5f2c81cad4c791d2f1f878b1379554" }
-#zksync_os_interface = { path = "../zksync-os-interface/crates/interface" }
+zksync_os_interface = { git = "https://github.com/matter-labs/zksync-os-interface", rev = "2fcc13cb594e08930d848dc6b610d180016db71b", default-features = false }
+#zksync_os_interface = { path = "../zksync-os-interface/crates/interface", default-features = false }
+
+zksync_os_types = { git = "https://github.com/matter-labs/zksync-os-interface", rev = "2fcc13cb594e08930d848dc6b610d180016db71b", default-features = false }
+#zksync_os_types = { path = "../zksync-os-interface/crates/types", default-features = false }
 
 risc_v_simulator = { git = "https://github.com/matter-labs/zksync-airbender", tag = "v0.4.3"}
 blake2s_u32 = { git = "https://github.com/matter-labs/zksync-airbender", tag = "v0.4.3"}
