+++ basic_bootloader/Cargo.toml
@@ -15,11 +15,10 @@ cycle_marker = { path = "../cycle_marker" }
 evm_interpreter = { path = "../evm_interpreter", default-features = false }
 system_hooks = { path = "../system_hooks", default-features = false }
 basic_system = { path = "../basic_system", default-features = false }
-crypto = { path = "../crypto", default-features = false }
-zksync_os_interface = { workspace = true }
-
 ruint = { version = "1.12.3", default-features = false, features = ["alloc"] }
-alloy = { version = "1", default-features = false }
+crypto = { path = "../crypto", default-features = false }
+zksync_os_types = { workspace = true, default-features = false }
+
 hex = {version = "*", optional = true}
 arrayvec = { version = "0.7.4", default-features = false }
 strum_macros = "0.27.1"
