+++ basic_system/Cargo.toml
@@ -16,7 +16,7 @@ crypto = { path = "../crypto", default-features = false, features = ["secp256k1-
 storage_models = { path = "../storage_models", default-features = false }
 modexp = { path = "../supporting_crates/modexp" }
 cycle_marker = { path = "../cycle_marker" }
-zksync_os_interface = { workspace = true }
+zksync_os_types = { workspace = true, default-features = false }
 
 ruint = { version = "1.12.3", default-features = false, features = ["alloc"] }
 arrayvec = { version = "0.7.4", default-features = false }
@@ -25,7 +25,7 @@ either = { version = "*", default-features = false }
 cfg-if = "1.0.0"
 const_for = "0.1.5"
 
-rand = {version = "0.9.0", optional = true}
+rand = { version = "0.9.0", optional = true }
 serde = { version = "1.0", default-features = false, features = ["derive", "alloc"], optional = true }
 strum_macros = "0.27.1"
 paste = "1.0.15"
@@ -36,7 +36,7 @@ evm-compatibility = []
 default = ["testing"]
 cycle_marker = ["cycle_marker/log_to_file"]
 wrap-in-batch = []
-proving = ["crypto/proving"]
+proving = ["crypto/proving", "zksync_os_types/portable-atomic"]
 evm_refunds = ["storage_models/evm_refunds"]
 # The finish method will just return inputs, so outputs can be procesed outside bootloader
 # Used for proving multi block batch within 1 run
