+++ tests/instances/eth_runner/src/post_check.rs
@@ -2,7 +2,7 @@ use crate::prestate::*;
 use crate::receipts::TransactionReceipt;
 use alloy::hex;
 use rig::log::{error, info};
-use rig::zksync_os_interface::types::BlockOutput;
+use rig::zksync_os_interface::output::BlockOutput;
 use ruint::aliases::{B160, B256, U256};
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
